requirements.txt
numpy
scikit-learn
pandas

torch>=1.4.0
torchvision
tensorflow
pillow

pyyaml
lmdb

tqdm
tabulate
matplotlib>=3.1.2
seaborn
